requirements.txt
djangorestframework
django-extensions
requests